requirements.txt
adafruit-python-shell
Cython==0.29.36
pyliblo3==0.12.0
RPi.GPIO==0.7.1
gpiozero
vl53l5cx-ctypes
numpy==1.26.4
pillow==9.1.0
Adafruit_GPIO
Adafruit-SSD1306